+++ packages/playground/src/ui.ts
@@ -1,1 +1,4 @@
+import './critical.css';
+import './playground.css';
+
 export * from './index';